requirements.txt
django
amqp==5.2.0;
APScheduler==3.10.4;
asgiref==3.7.2;
billiard==4.2.0;
CacheControl==0.13.1;
cachetools==5.3.2;
celery==5.3.6;
certifi==2023.11.17;
cffi==1.16.0;
charset-normalizer==3.3.2;
click==8.1.7;
click-didyoumean==0.3.0;
click-plugins==1.1.1;
click-repl==0.3.0; 
cryptography==41.0.7;
Cython==3.0.7;
decorator==4.4.2;
Django==4.2.7;
django-background-tasks==1.2.5;
django-compat==1.0.15;
django-crontab==0.7.1;
djangorestframework==3.14.0;
filelock==3.13.1;
firebase-admin==6.3.0;
fsspec==2023.12.2;
geographiclib==2.0;
geopy==2.4.1;
google-api-core==2.15.0;
google-api-python-client==2.113.0;
google-auth==2.26.1;
google-auth-httplib2==0.2.0;
google-cloud-core==2.4.1;
google-cloud-firestore==2.14.0;
google-cloud-storage==2.14.0;
google-crc32c==1.5.0;           
google-resumable-media==2.7.0;             
googleapis-common-protos==1.62.0;
grpcio==1.60.0;
grpcio-status==1.60.0;
httplib2==0.22.0;
idna==3.6;
Jinja2==3.1.2;
joblib==1.3.2;
kombu==5.3.4;
MarkupSafe==2.1.3;
mpmath==1.3.0;
msgpack==1.0.7;
networkx==3.2.1;
numpy==1.26.2;
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12>=2.21.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
pandas==1.3.5
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
torch-geometric-temporal==0.54.0
torch_geometric==2.4.0
tqdm==4.66.1
# triton==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
# pandas==1.3.5
# prompt-toolkit==3.0.43
# proto-plus==1.23.0
# protobuf==4.25.1
# psutil==5.9.7
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pycparser==2.21
# PyJWT==2.8.0
# pyparsing==3.1.1
# python-dateutil==2.8.2
# pytz==2023.3.post1
# requests==2.31.0
# rsa==4.9
# scikit-learn==1.3.2
# scipy==1.11.4
# six==1.16.0
# sqlparse==0.4.4
# sympy==1.12
# threadpoolctl==3.2.0
# torch-geometric-temporal==0.54.0
# torch_geometric==2.4.0
# tqdm==4.66.1
# triton
# typing_extensions==4.8.0
# tzdata==2023.3
# tzlocal==5.2
# uritemplate==4.1.1
# urllib3==2.1.0
# vine==5.1.0
# wcwidth==0.2.13
# nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
pandas==1.3.5;
prompt-toolkit==3.0.43;
proto-plus==1.23.0;
protobuf==4.25.1;
psutil==5.9.7;
pyasn1==0.5.1;
pyasn1-modules==0.3.0;
pycparser==2.21;
PyJWT==2.8.0;
pyparsing==3.1.1;
python-dateutil==2.8.2;
pytz==2023.3.post1;
requests==2.31.0;
rsa==4.9;
scikit-learn==1.3.2;
scipy==1.11.4;
six==1.16.0;
sqlparse==0.4.4;
sympy==1.12;
threadpoolctl==3.2.0;
torch-geometric-temporal;
torch_geometric==2.4.0;
tqdm==4.66.1;
typing_extensions==4.8.0;
tzdata==2023.3;
tzlocal==5.2
uritemplate==4.1.1;
urllib3==2.1.0;
vine==5.1.0;
wcwidth==0.2.13;
# https://huggingface.co/r4ziel/xformers_pre_built/resolve/main/triton-2.0.0-cp310-cp310-win_amd64.whl
django-cors-headers
# https://huggingface.co/r4ziel/xformers_pre_built/resolve/main/triton-2.0.0-cp310-cp310-win_amd64.whl
